fix: moved interfaces, button fixes
