+++ src/components/Button/Button.tsx
@@ -1,48 +1,26 @@
 import React from 'react';
 import {Text, TouchableOpacity} from 'react-native';
-import {ButtonI} from '../../interfaces/ButtonInterface';
+import {ButtonI} from './ButtonInterface';
 import styles from './ButtonStyles';
 
 export const Button = ({
   disabled = false,
-  bgColor = '#0D0113',
-  borderColor = '#0D0113',
   text = 'Button',
-  textColor = '#fff',
+  buttonContainerStyle = {},
+  buttonTextStyle = {},
   handlePress,
 }: ButtonI) => {
   return (
     <TouchableOpacity
       activeOpacity={0.8}
       onPress={() => {
-        // console.log("Button clicked!");
         if (!disabled && handlePress) {
           handlePress();
         }
       }}
-      style={[
-        styles.buttonWrapper,
-        {
-          backgroundColor:
-            bgColor && !disabled ? bgColor : 'rgba(13, 1, 19, 0.2)',
-          borderColor:
-            borderColor && !disabled
-              ? borderColor
-              : disabled
-              ? 'transparent'
-              : '#0D0113',
-        },
-      ]}
+      style={[styles.buttonWrapper, buttonContainerStyle]}
       disabled={disabled}>
-      <Text
-        style={[
-          styles.buttonText,
-          {
-            color: textColor ?? '#fff',
-          },
-        ]}>
-        {text}
-      </Text>
+      <Text style={[styles.buttonText, buttonTextStyle]}>{text}</Text>
     </TouchableOpacity>
   );
 };
