+++ src/components/Button/GradientButton.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Text, TouchableOpacity} from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
-import {GradientButtonI} from '../../interfaces/ButtonInterface';
+import {GradientButtonI} from './ButtonInterface';
 import styles from './ButtonStyles';
 
 export const GradientButton = ({
