+++ src/components/CustomImage/CustomImageInterface.ts
