+++ src/components/SwipeToDelete/SwipeToDeleteInterface.ts
