+++ src/components/Switch/SwitchInterface.ts
