+++ src/components/Tabs/TabsInterface.ts
